Split moderation playlist state
